слишком скучно

+++ src/main/java/org/flamierawieo/x00FA9A/client/input/MouseInputListener.java
@@ -2,7 +2,7 @@ package org.flamierawieo.x00FA9A.client.input;
 
 public interface MouseInputListener {
 
-    void onMouseMove(double x, double y);
+    void onMouseMove(float x, float y);
     void onMouseButtonDown(int button, int mods);
     void onMouseButtonUp(int button, int mods);
