+++ src/main/java/org/flamierawieo/x00FA9A/shared/Tickable.java
@@ -2,6 +2,6 @@ package org.flamierawieo.x00FA9A.shared;
 
 public interface Tickable {
 
-    void tick(double delta);
+    void tick(float delta);
 
 }
